Fix a wrapped match distance, harden the FFI and gzip write paths, and strengthen the test surface
Deflate
- Widen `DeflateState::prev_match` from `u16` to `usize` so it is as wide as
  `match_start`, which C declares as `IPos`. `fill_window` slides the window by
  subtracting `w_size` from both `strstart` and `match_start`, and a stale
  `match_start` is allowed to wrap; because C evaluates `strstart - 1 -
  prev_match` in the same modular unsigned arithmetic, the wrap cancels exactly
  and the true distance falls out. Narrowing to 16 bits destroyed that
  cancellation, so a wrapped value produced a bogus distance: an overflow panic
  in `deflate_slow` under debug assertions and an out-of-bounds `d_code` index
  in release, both reachable from the safe public API. The deferred-match
  emission now copies `match_start` at full width and computes the distance with
  wrapping arithmetic. Covered by a regression test that reproduces the wrap
  (`strstart = 250`, `prev_match = 2^64-1`, `prev_length = 7`, `w_size = 512`)
  and pins the output to reference C zlib's 943 bytes and CRC-32 `0xb573cc1b`.

C ABI boundary
- `deflateParams` no longer builds slices from `next_in`/`next_out` unless the
  pre-flush will actually run. Constructing a Rust slice asserts that the whole
  declared region is one live allocation even when no byte is touched, so
  building them unconditionally turned a call reference zlib answers with `Z_OK`
  into undefined behaviour for a caller whose buffers are stale or undersized.
- Document the defined rejections of the deflate and inflate entry points: a
  null `z_streamp`, a zeroed or already-ended stream, a mismatched handle kind,
  and inconsistent `next_*`/`avail_*` pairs.
- Hide the no-std `rust_eh_personality` stub with a `global_asm!` `.hidden`
  (`.private_extern` on Mach-O). The symbol is still defined exactly once, so
  the freestanding artifacts link, but it is no longer exported: every feature
  row now presents the same 95-symbol dynamic surface, and `dlsym` for it
  returns null.

gzip file I/O
- Track an `out_start` cursor mirroring C's `state->x.next` so a short write
  advances through the pending window instead of copying the remainder to the
  front, removing quadratic behaviour on a destination that accepts little at a
  time.
- Credit `pos` and `skip` with the bytes the engine actually ingested before
  propagating a failure, matching `gz_zero`'s C ordering, and stop rather than
  spin if a step reports success while consuming nothing.
- Release `in_buf` and `out_buf` on every initialisation failure path before
  recording the error, so a failed open leaks nothing and allocation timing
  still matches C.

Benchmarks
- Import `compress_bound`, `compress2` and `uncompress` from the crate root
  rather than through internal module paths.
- Validate every case's output before it is timed: compress, decode, and assert
  both the exact recovered length and byte equality, outside the timed closure.
- Add an incompressible-input profile at levels 1, 6 and 9, and correct the
  documented throughput figures and their references.
- Replace the scalar-versus-SIMD comparison with two rows that genuinely differ
  in the `simd` feature, and record why adding `--features simd` on top of the
  defaults compares nothing.

Fuzz harnesses
- Replace permissive return-code checks with per-entry-point contract sets
  derived from `zlib.h` and cross-checked against the implementation, so an
  out-of-contract answer fails loudly instead of passing silently.
- Remove the bail-outs that let a target skip its own oracle, bound output
  growth and pass counts so a stall is reported rather than spun on, and gate
  gzip-dependent items individually so both feature rows build.
- Deny undocumented and multi-operation unsafe blocks in the fuzz manifest and
  bring every site into line.

Tests and tooling
- Add a fault-injection seam to the gzip write path, compiled only under
  `cfg(test)`, and cover interrupted writes, stalls, and a destination that
  accepts nothing, in both the buffered and direct paths.
- Create temporary files exclusively inside a private mode-0700 directory
  behind an RAII guard instead of predictable shared paths.
- Trim the lint, format, toolchain and dependency-policy files to settings that
  are actually in force, and correct the surrounding documentation.

+++ src/deflate/state.rs
@@ -657,7 +657,25 @@ pub struct DeflateState {
     /// Length of the best match found (C `match_length`).
     pub match_length: usize,
     /// Previous match position (C `prev_match`, an `IPos`).
-    pub prev_match: u16,
+    ///
+    /// Held at the **same width as [`match_start`](Self::match_start)**, which
+    /// it is a verbatim copy of (`deflate.c` L1985: `s->prev_match =
+    /// s->match_start`). C declares `prev_match` as `IPos` (`typedef unsigned
+    /// IPos`, `deflate.h` L98) — exactly as wide as the `uInt match_start`
+    /// (`deflate.h` L167) it receives — and that equal width is load-bearing,
+    /// not incidental.
+    ///
+    /// The window slide in [`fill_window`](Self::fill_window) subtracts
+    /// `w_size` from `match_start` unconditionally (`deflate.c` L288), so a
+    /// `match_start` that is stale rather than freshly set by
+    /// [`longest_match`](Self::longest_match) wraps around. C recovers the true
+    /// distance regardless, because `strstart` and `match_start` are reduced by
+    /// the same amount on every slide and `strstart - 1 - prev_match` is
+    /// evaluated in the same modular arithmetic — the wrap cancels exactly.
+    /// Narrowing this field would discard the high bits of a wrapped value,
+    /// break that cancellation, and yield a bogus match distance (see
+    /// `deflate::slow::deflate_slow`).
+    pub prev_match: usize,
     /// Set when a deferred (lazy) match from the previous step exists (C
     /// `match_available`, an `int` used as a boolean).
     pub match_available: bool,
@@ -2471,14 +2489,14 @@ mod tests {
     /// the crate's single authoritative tuning table, and there must be exactly
     /// one such table.
     ///
-    /// `src/deflate/state.rs` previously carried a private duplicate of
-    /// `configuration_table`, so `lm_init` (init) and `deflate_params`
-    /// (re-dispatch) read *different* copies of the same numbers. Two copies can
-    /// diverge silently, and any divergence changes the lazy-match decisions and
+    /// The invariant matters because two copies of the tuning numbers can diverge
+    /// silently: if `lm_init` (init) and `deflate_params` (re-dispatch) were to read
+    /// *different* tables, any divergence would change the lazy-match decisions and
     /// therefore the emitted token stream — breaking byte-identity with reference
-    /// zlib without breaking decodability (AAP §0.6.4, TO-5). The duplicate is
-    /// gone; this test pins the surviving one against live state for all ten
-    /// levels plus the `Z_DEFAULT_COMPRESSION` alias.
+    /// zlib without breaking decodability (AAP §0.6.4, TO-5). There is exactly one
+    /// table, [`crate::deflate::strategy::CONFIGURATION_TABLE`], and this test pins
+    /// it against live state for all ten levels plus the `Z_DEFAULT_COMPRESSION`
+    /// alias.
     #[test]
     fn lm_init_loads_every_level_from_the_authoritative_table() {
         for level in 0..=9i32 {
@@ -2746,9 +2764,10 @@ mod tests {
     /// `deflateCopy` produces a genuinely independent snapshot of a **live**
     /// compressor, not a shallow alias.
     ///
-    /// The previous version of this test cloned a *freshly initialised* state,
-    /// poked one window byte, and compared lengths — which a shallow copy that
-    /// shared every buffer would also have passed. This version:
+    /// Cloning a *freshly initialised* state, poking one window byte and comparing
+    /// lengths would be satisfied by a shallow copy that shared every buffer. The
+    /// snapshot is therefore taken mid-compression and the independence asserted
+    /// directly:
     ///
     /// 1. drives real compression so the sliding window, the `head`/`prev` hash
     ///    chains, the overlaid symbol region and the pending buffer all hold live
@@ -3629,7 +3648,7 @@ mod tests {
     /// Both are offsets from the same base — `sym_buf` is `pending_buf +
     /// lit_bufsize` — so subtracting it leaves
     /// `lit_bufsize < pending_out + ((Buf_size + 7) >> 3)`, which is the
-    /// expression `deflate_prime` evaluates verbatim now that both quantities are
+    /// expression `deflate_prime` evaluates verbatim, because both quantities are
     /// stored as indices. This test pins the boundary from both sides at the
     /// smallest `lit_bufsize` (`mem_level = 1`, 128 bytes), where the guard is
     /// tightest.
